Update board rendering after a loss
Render bombs is no longer used only for debugging.
It is now used to show the player where all the bombs were as well as
whether or not they correctly placed their flags.

render_bombs now also uses Tile.tile_string in order to apply DRY
principles and increase consistency in how the board is rendered.